% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpmodels-package.R
\docType{package}
\name{bpmodels-package}
\alias{bpmodels}
\alias{bpmodels-package}
\title{bpmodels: Analysing chain statistics using branching process models}
\description{
Provides methods to analyse and simulate the size and length of branching processes with an arbitrary offspring distribution. These can be used, for example, to analyse the distribution of chain sizes or length of infectious disease outbreaks, as discussed in Farrington et al. (2003) \doi{10.1093/biostatistics/4.2.279}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sbfnk/bpmodels}
  \item Report bugs at \url{https://github.com/sbfnk/bpmodels/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}

Authors:
\itemize{
  \item Flavio Finger \email{flavio.finger@epicentre.msf.org}
}

Other contributors:
\itemize{
  \item Zhian N. Kamvar \email{zkamvar@gmail.com} [contributor]
}

}
\keyword{internal}
