% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{chain_sim}
\alias{chain_sim}
\title{Simulate chains using a branching process}
\usage{
chain_sim(n, offspring, stat = c("size", "length"), infinite = Inf,
  tree = FALSE, ...)
}
\arguments{
\item{n}{number of simulations to run.}

\item{offspring}{offspring distribution: a character string corresponding to
the R distribution function (e.g., "pois" for Poisson, where
\code{\link{rpois}} is the R function to generate Poisson random numbers)}

\item{stat}{statistic to calculate ("size" or "length" of chains)}

\item{infinite}{a size or length from which the size/length is to be
considered infinite}

\item{tree}{return the tree of infectors}

\item{...}{parameters of the offspring distribution}
}
\value{
a vector of sizes/lengths (if \code{tree==FALSE}), or a data frame
    with columns `n` (simulation ID), `id` (a unique ID within each
    simulation for each individual element of the chain), `ancestor` (the ID
    of the ancestor of each element) and `generation`.
}
\description{
Simulate chains using a branching process
}
\examples{
chain_sim(n=5, "pois", "size", lambda=0.5)
}
\author{
Sebastian Funk
}
