% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_susceptibles.R
\name{chain_sim_susc}
\alias{chain_sim_susc}
\title{Simulate a single chain using a branching process while accounting
for depletion of susceptibles.}
\usage{
chain_sim_susc(
  offspring = c("pois", "nbinom"),
  mn_offspring,
  disp_offspring,
  serial,
  t0 = 0,
  tf = Inf,
  pop,
  initial_immune = 0
)
}
\arguments{
\item{offspring}{offspring distribution: a character string corresponding to
the R distribution function. Currently only "pois" & "nbinom" are
supported. Internally truncated distributions are used to avoid infecting
more people than susceptibles available.}

\item{mn_offspring}{the average number of secondary cases for each case}

\item{disp_offspring}{the dispersion coefficient (var/mean) of the number of
secondary cases. Ignored if offspring == "pois". Must be > 1.}

\item{serial}{the serial interval. A function that takes one parameter
(\code{n}), the number of serial intervals to randomly sample.
Value must be >= 0.}

\item{t0}{start time}

\item{tf}{end time}

\item{pop}{the population}

\item{initial_immune}{the number of initial immunes in the population}
}
\value{
a data frame with columns \code{time}, \code{id} (a unique ID for each
individual element of the chain), \code{ancestor} (the ID of the ancestor
of each element), and \code{generation}.
}
\description{
Simulate a single chain using a branching process while accounting
for depletion of susceptibles.
}
\details{
This function has a couple of key differences with chain_sim:
it can only simulate one chain at a time,
it can only handle implemented offspring distributions
("pois" and "nbinom"),
it always tracks and returns a data frame containing the entire tree,
the maximal length of chains is limited with pop instead of infinite.
}
\examples{
chain_sim_susc("pois", mn_offspring=0.5, serial = function(x) 3, pop = 100)
}
\author{
Flavio Finger
}
