% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid19_sa}
\alias{covid19_sa}
\title{COVID-19 Data Repository for South Africa}
\format{
\subsection{\code{covid19_sa}}{

A data frame with 19 rows and 2 columns:
\describe{
\item{date}{Date case was reported}
\item{cases}{Number of cases}
...
}
}
}
\source{
\url{https://github.com/dsfsi/covid19za}
Further details in \code{data-raw/covid19_sa.R}.
}
\usage{
covid19_sa
}
\description{
An aggregated subset of the COVID-19 Data Repository for South Africa created,
maintained and hosted by Data Science for Social Impact research group,
led by Dr. Vukosi Marivate ...
}
\details{
The data is originally provided as a linelist but has been subsetted and
cleaned in \code{data-raw/covid19_sa.R}.
}
\keyword{datasets}
