% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codes.R
\name{envelope.BP}
\alias{envelope.BP}
\title{Envelope BP}
\usage{
envelope.BP(model, k = 100, link = c("log", "identity"),
  color = "grey50", xlabel = "Theorical Quantile",
  ylabel = "Empirical Quantile", font = "serif")
}
\description{
Envelope BP
}
