% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codes.R
\name{envelope.GA}
\alias{envelope.GA}
\title{Envelope GA}
\usage{
envelope.GA(model, k = 100, link = c("log", "identity"),
  color = "grey50", xlabel = "Theorical Quantile",
  ylabel = "Empirical Quantile", font = "serif")
}
\description{
Envelope GA
}
