% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codes.R
\name{envelope}
\alias{envelope}
\title{Envelope BS}
\usage{
envelope(model, k = 100, precision = c("fixed", "varying"),
  dist = RBS(mu.link = "identity", sigma.link = "identity"),
  color = "grey50", xlabel = "Theorical quantile",
  ylabel = "Empirical quantile", font = "serif")
}
\description{
Envelope BS
}
