% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{find_function_name}
\alias{find_function_name}
\title{Finds the name of a function passed as an argument}
\usage{
find_function_name(fun)
}
\arguments{
\item{fun}{function of which the name is to be determined}
}
\value{
function name
}
\description{
This works even when a function is passed multiple times (e.g., when used
inside an \code{\link{optim}} call).
See https://stackoverflow.com/a/46740314/10886760
}
\author{
Sebastian Funk
}
\keyword{internal}
