% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{geom_length_ll}
\alias{geom_length_ll}
\title{Log-likelihood of the length of chains with geometric offspring distribution}
\usage{
geom_length_ll(x, prob)
}
\arguments{
\item{x}{vector of sizes}

\item{prob}{probability of the geometric distribution with mean
\code{1/prob}}
}
\value{
log-likelihood values
}
\description{
Log-likelihood of the length of chains with geometric offspring distribution
}
\author{
Sebastian Funk
}
\keyword{internal}
