% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{offspring_ll}
\alias{offspring_ll}
\title{Likelihood of the length of chains with generic offspring distribution}
\usage{
offspring_ll(x, offspring, stat, nsim_offspring = 100, ...)
}
\arguments{
\item{x}{vector of sizes}

\item{offspring}{Offspring distribution: a character string corresponding to
the R distribution function (e.g., "pois" for Poisson, where
\code{\link{rpois}} is the R function to generate Poisson random numbers)}

\item{stat}{statistic given as \code{x} ("size" or "length" of chains)}

\item{nsim_offspring}{number of simulations of the offspring distribution
for approximation the size/length distribution}

\item{...}{any parameters to pass to \code{\link{chain_sim}}}
}
\value{
log-likelihood values
}
\description{
The likelihoods are calculated with a crude approximation using simulated
chains by linearly approximating any missing values in the empirical
cumulative distribution function (ecdf).
}
\author{
Sebastian Funk
}
\keyword{internal}
