% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{pois_length_ll}
\alias{pois_length_ll}
\title{Log-likelihood of the length of chains with Poisson offspring distribution}
\usage{
pois_length_ll(x, lambda)
}
\arguments{
\item{x}{vector of sizes}

\item{lambda}{rate of the Poisson distribution}
}
\value{
log-likelihood values
}
\description{
Log-likelihood of the length of chains with Poisson offspring distribution
}
\author{
Sebastian Funk
}
\keyword{internal}
