% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{rbinom_size}
\alias{rbinom_size}
\title{Samples size (the number of trials) of a binomial distribution}
\usage{
rbinom_size(n, x, prob)
}
\arguments{
\item{n}{number of samples to generate##'      secondary cases. Ignored if offspring == "pois". Must be > 1.}

\item{x}{number of successes}

\item{prob}{probability of success}

\item{serial}{the serial interval. A function that takes one parameter
(`n`), the number of serial intervals to randomly sample.
Value must be >= 0.}

\item{t0}{start time}

\item{tf}{end time}

\item{pop}{the population}
}
\value{
sampled sizes
}
\description{
Samples the size parameter from the binomial distribution with fixed x
(number of successes) and p (success probability)
}
\author{
Sebastian Funk
}
\keyword{internal}
