% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borel.r
\name{rborel}
\alias{rborel}
\title{Random number generator from the Borel distribution}
\usage{
rborel(n, mu, infinite = Inf)
}
\arguments{
\item{n}{number of random variates to generate.}

\item{mu}{mu parameter (the Poisson mean).}

\item{infinite}{any number to treat as infinite; simulations will be stopped
if this number is reached}
}
\value{
vector of random numbers
}
\description{
Random numbers are generated by simulating from a Poisson branching process
}
\author{
Sebastian Funk
}
