% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{rgen_length}
\alias{rgen_length}
\title{Samples chain lengths with given observation probabilities}
\usage{
rgen_length(n, x, prob)
}
\arguments{
\item{n}{number of samples to generate}

\item{x}{observed chain lengths}

\item{prob}{probability of observation}
}
\value{
sampled lengths
}
\description{
Samples the length of a transmission chain where each individual element is
observed with binomial probability with parameters n (number of successes)
and p (success probability)
}
\author{
Sebastian Funk
}
\keyword{internal}
