% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{rnbinom_mean_disp}
\alias{rnbinom_mean_disp}
\title{Negative binomial random numbers parametrized
in terms of mean and dispersion coefficient}
\usage{
rnbinom_mean_disp(n, mn, disp)
}
\arguments{
\item{n}{number of samples to draw}

\item{mn}{mean of distribution; Must be > 0.}

\item{disp}{dispersion coefficient (var/mean); Must be > 1.}
}
\value{
vector containing the random numbers
}
\description{
Negative binomial random numbers parametrized
in terms of mean and dispersion coefficient
}
\examples{
rnbinom_mean_disp(n = 5, mn = 4, disp = 2)
}
\author{
Flavio Finger
}
