% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codes.R
\name{sim.gee.bp}
\alias{sim.gee.bp}
\title{Simulations}
\usage{
sim.gee.bp(n=10,r=2,phi=2,REP=10,alpha=0.5)
}
\arguments{
\item{n}{sampling size}

\item{r}{xx}

\item{phi}{xx}

\item{REP}{xx}

\item{alpha}{xx}
}
\description{
Simulations
}
