% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpmodels-package.R
\docType{package}
\name{bpmodels-package}
\alias{bpmodels}
\alias{bpmodels-package}
\title{bpmodels: Analysing Transmission Chain Statistics using Branching Process Models}
\description{
Provides methods to analyse and simulate the size and length of branching processes with an arbitrary offspring distribution. These can be used, for example, to analyse the distribution of chain sizes or length of infectious disease outbreaks, as discussed in Farrington et al. (2003) \doi{10.1093/biostatistics/4.2.279}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/bpmodels}
  \item \url{https://epiverse-trace.github.io/bpmodels/}
  \item Report bugs at \url{https://github.com/epiverse-trace/bpmodels/issues}
}

}
\author{
\strong{Maintainer}: James M. Azam \email{james.azam@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5782-7330}{ORCID})

Authors:
\itemize{
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-2842-3406}{ORCID}) [copyright holder]
  \item Flavio Finger \email{flavio.finger@epicentre.msf.org} (\href{https://orcid.org/0000-0002-8613-5170}{ORCID})
}

Other contributors:
\itemize{
  \item Zhian N. Kamvar \email{zkamvar@gmail.com} (\href{https://orcid.org/0000-0003-1458-7108}{ORCID}) [contributor]
}

}
\keyword{internal}
