% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{chain_ll}
\alias{chain_ll}
\title{Likelihood for the outcome of a branching process}
\usage{
chain_ll(x, offspring, ..., obs_prob = 1, stat = c("size", "length"),
  infinite = Inf, exclude)
}
\arguments{
\item{x}{vector of sizes or lengths of transmission chains}

\item{offspring}{offspring distribution as character string, e.g. "pois" for
the Poisson offspring distribution.}

\item{...}{parameters for the offspring distribution}

\item{obs_prob}{observation probability (assumed constant)}

\item{stat}{statistic given as \code{x} ("size" or "length" of chains)}

\item{infinite}{any chains of this size/length will be treated as infinite}

\item{exclude}{any sizes/lengths to exclude from the likelihood calculation}
}
\value{
likelihood
}
\description{
Likelihood for the outcome of a branching process
}
\seealso{
pois_size_ll nbinom_size_ll gborel_size_ll pois_length_ll geom_length_ll offspring_ll
}
\author{
Sebastian Funk
}
