% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{chain_sim}
\alias{chain_sim}
\title{Simulate chains using a branching process}
\usage{
chain_sim(n, offspring, stat = c("size", "length"), infinite = Inf,
  ...)
}
\arguments{
\item{n}{number of simulations to run.}

\item{offspring}{offspring distribution, given as the function used to
generate the number of offspring in each generation, e.g. `rpois` for
Poisson distributed offspring}

\item{stat}{statistic to calculate ("size" or "length" of chains)}

\item{infinite}{a size or length from which the size/length is to be
considered infinite}

\item{...}{parameters of the offspring distribution}
}
\value{
a vector of sizes/lengths
}
\description{
Simulate chains using a branching process
}
\examples{
chain_sim(n=5, rpois, "size", lambda=0.5)
}
\author{
Sebastian Funk
}
