% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_susceptibles.R
\name{chain_sim_susc}
\alias{chain_sim_susc}
\title{Simulate a single chain using a branching process while accounting
for depletion of susceptibles.}
\usage{
chain_sim_susc(offspring = c("pois", "nbinom"), mn_offspring,
  disp_offspring = NULL, serial, t0 = 0, tf = Inf, pop,
  initial_immune = 0, adjust_params = NULL)
}
\arguments{
\item{offspring}{offspring distribution: a character string corresponding to
the R distribution function. Currently only "pois" & "nbinom" are
supported. Internally truncated distributions are used to avoid infecting
more people than susceptibles available.}

\item{mn_offspring}{the average number of secondary cases for each case}

\item{disp_offspring}{the dispersion coefficient (var/mean) of the number of
secondary cases. Ignored if offspring == "pois". Must be > 1.}

\item{serial}{the serial interval. A function that takes one parameter
(`n`), the number of serial intervals to randomly sample.
Value must be >= 0.}

\item{t0}{start time}

\item{tf}{end time}

\item{pop}{the population}

\item{initial_immune}{the number of initial immunes in the population}

\item{adjust_params}{a function that takes the current time t and simulation
parameters (susc, mn_offspring, disp_offspring), modifies them and
returns them as a list. Modifications to susc are kept until the end of
the chain, modifications to mn and disp are only for the current time t}
}
\value{
a data frame with columns `time`, `id` (a unique ID for each
    individual element of the chain), `ancestor` (the ID of the ancestor
     of each element), and `generation`.
}
\description{
Simulate a single chain using a branching process while accounting
for depletion of susceptibles.
}
\details{
This function has a couple of key differences with chain_sim:
    it can only simulate one chain at a time,
    it can only handle implemented offspring distributions
        ("pois" and "nbinom"),
    it always tracks and returns a data frame containing the entire tree,
    the maximal length of chains is limited with pop instead of infinite.
}
\examples{
chain_sim_susc("pois", mn_offspring=0.5, serial = function(x) 3, pop = 100)
}
\author{
Flavio Finger
}
