% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borel.r
\name{dborel}
\alias{dborel}
\title{Density of the Borel distribution}
\usage{
dborel(x, mu, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles; integer.}

\item{mu}{mu parameter (the poisson mean); non-negative.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
probability mass.
}
\description{
Density of the Borel distribution
}
\author{
Sebastian Funk
}
