% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{rbinom_size}
\alias{rbinom_size}
\title{Samples size (the number of trials) of a binomial distribution}
\usage{
rbinom_size(n, x, prob)
}
\arguments{
\item{n}{number of samples to generate}

\item{x}{number of successes}

\item{prob}{probability of success}
}
\value{
sampled sizes
}
\description{
Samples the size parameter from the binomial distribution with fixed x
(number of successes) and p (success probability)
}
\author{
Sebastian Funk
}
\keyword{internal}
